FWMT-289 Code review fixes
