+++ src/main/java/uk/gov/ons/fwmt/staff_service/data/dto/UserDTO.java
@@ -1,8 +1,14 @@
 package uk.gov.ons.fwmt.staff_service.data.dto;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
 public class UserDTO {
 
     public String authNo;
